TD-1220:Review comments implemented
